test(ios): extract snapshot truncation regressions
