+++ src/__tests__/client-transport-fixture.ts
@@ -1,0 +1,36 @@
+import type { AgentDeviceClientConfig } from '../agent-device-client.ts';
+import type { DaemonRequest, DaemonResponse } from '@agent-device/kernel/contracts';
+import { mkdtempForTestSync } from './test-utils/tmp-dir.ts';
+
+const TEST_STATE_DIR = mkdtempForTestSync('agent-device-client-test-');
+
+export function createTransport(
+  handler: (req: Omit<DaemonRequest, 'token'>) => Promise<DaemonResponse> | DaemonResponse,
+): {
+  calls: Array<Omit<DaemonRequest, 'token'>>;
+  config: AgentDeviceClientConfig;
+  transport: (req: Omit<DaemonRequest, 'token'>) => Promise<DaemonResponse>;
+} {
+  const calls: Array<Omit<DaemonRequest, 'token'>> = [];
+  const config: AgentDeviceClientConfig = {
+    session: 'qa',
+    stateDir: TEST_STATE_DIR,
+    cwd: '/tmp/agent-device',
+    debug: true,
+    daemonBaseUrl: 'http://daemon.example.test',
+    daemonAuthToken: 'secret',
+    daemonTransport: 'http',
+    tenant: 'acme',
+    sessionIsolation: 'tenant',
+    runId: 'run-123',
+    leaseId: 'lease-123',
+  };
+  return {
+    calls,
+    config,
+    transport: async (req) => {
+      calls.push(req);
+      return await handler(req);
+    },
+  };
+}
