+++ src/daemon/handlers/__tests__/snapshot-handler.test.ts
@@ -6,22 +6,18 @@ import {
   resetGetRuntimeFixture,
 } from './interaction-get-runtime-fixture.ts';
 import fs from 'node:fs';
-import path from 'node:path';
 import { handleSnapshotCommands as handleProductionSnapshotCommands } from '../snapshot.ts';
 import { captureSnapshot } from '../snapshot-capture.ts';
 import { SessionStore } from '../../session-store.ts';
 import { setActiveProviderDeviceRuntimes } from '../../../provider-device-runtime.ts';
-import type { ProviderDeviceRuntime } from '@agent-device/contracts/device';
 import type { DaemonResponse, SessionState } from '../../types.ts';
 import { AppError } from '@agent-device/kernel/errors';
 import { platformResourceCleanup } from '../../../platform-runtime-resource-cleanup.ts';
 import { buildSnapshotSignatures } from '../../../snapshot/snapshot-freshness/index.ts';
 import { buildInteractionSurfaceSignature } from '../../interaction-outcome-policy.ts';
 import { buildSnapshotPresentationKey } from '@agent-device/kernel/snapshot';
-import { attachSnapshotPresentationEvidence } from '@agent-device/contracts/capture';
 import { snapshotCliOutput } from '../../../commands/capture/output.ts';
 import type { CaptureSnapshotResult } from '@agent-device/contracts/client';
-import { mkdtempForTestSync } from '../../../__tests__/test-utils/tmp-dir.ts';
 import {
   fixtureScreenshotCaptures,
   fixtureSettingsMutations,
@@ -29,6 +25,11 @@ import {
   snapshotRuntimeFixture,
 } from '../../__tests__/snapshot-runtime-fixture.ts';
 import type { BindDeviceRuntime } from '../../request-runtime-binding.ts';
+import {
+  makeProviderRuntimeOwning,
+  makeSession,
+  makeSessionStore,
+} from './snapshot-handler-fixture.ts';
 
 vi.mock('../snapshot-interactor-capture.ts', async () => {
   const fixture = await import('../../__tests__/legacy-snapshot-capture-fixture.ts');
@@ -75,15 +76,6 @@ function handleSnapshotCommands(
   });
 }
 
-function makeSessionStore(): SessionStore {
-  const root = mkdtempForTestSync('agent-device-snapshot-handler-');
-  return new SessionStore(path.join(root, 'sessions'));
-}
-
-type SessionExtra = Partial<SessionState>;
-function makeSession(name: string, d: SessionState['device'], extra?: SessionExtra): SessionState {
-  return { name, device: d, createdAt: Date.now(), actions: [], ...extra };
-}
 // An Apple wait runs inside an opened app: that bundle id is XCUITest's attach target, and
 // without one the plan asks for the without-active-app row local Apple refuses.
 const appAttach = (d: SessionState['device']): Partial<SessionState> =>
@@ -124,20 +116,6 @@ const providerIosDevice: SessionState['device'] = {
   target: 'mobile',
   booted: true,
 };
-
-function makeProviderRuntimeOwning(
-  device: SessionState['device'],
-  provider = 'browserstack',
-): ProviderDeviceRuntime {
-  return {
-    provider,
-    leaseLifecycle: {},
-    deviceInventoryProvider: async () => [device],
-    ownsDevice: (candidate) => candidate.id === device.id,
-    getInteractor: () => undefined,
-    shutdown: async () => undefined,
-  };
-}
 
 afterEach(() => {
   setActiveProviderDeviceRuntimes([]);
@@ -486,55 +464,6 @@ test('snapshot on provider-backed iOS runs without a tracked app', async () => {
   // guaranteed-useless spawn on what is now a success path.
   expect(mockBuildIosOpenCommandHint).not.toHaveBeenCalled();
   expect(bindCount).toBe(1);
-});
-
-test('Limrun unknown truncation stays omitted through daemon and public output', async () => {
-  const sessionStore = makeSessionStore();
-  const sessionName = 'limrun-ios-unknown-truncation';
-  const limrunDevice: SessionState['device'] = {
-    platform: 'apple',
-    appleOs: 'ios',
-    id: 'limrun:ios:lease-a',
-    name: 'Limrun iOS',
-    kind: 'simulator',
-    target: 'mobile',
-    booted: true,
-  };
-  sessionStore.set(sessionName, makeSession(sessionName, limrunDevice));
-  setActiveProviderDeviceRuntimes([makeProviderRuntimeOwning(limrunDevice, 'limrun')]);
-  legacyDispatchCapture.mockResolvedValue(
-    attachSnapshotPresentationEvidence(
-      {
-        nodes: [{ index: 0, depth: 0, type: 'Application', label: 'Demo' }],
-        backend: 'xctest',
-        producer: 'limrun-ios-tree',
-        warnings: ['tree completeness is not independently verified'],
-      },
-      { owner: 'ios-snapshot-engine' },
-    ),
-  );
-
-  const response = await handleSnapshotCommands({
-    req: {
-      token: 't',
-      session: sessionName,
-      command: 'snapshot',
-      positionals: [],
-      flags: {},
-    },
-    sessionName,
-    logPath: '/tmp/daemon.log',
-    sessionStore,
-  });
-
-  expect(response?.ok).toBe(true);
-  if (!response?.ok) return;
-  expect(response.data).not.toHaveProperty('truncated');
-
-  const cliOutput = await snapshotCliOutput({
-    result: response.data as unknown as CaptureSnapshotResult,
-  });
-  expect(cliOutput.jsonData).not.toHaveProperty('truncated');
 });
 
 test('diff on local iOS still requires a tracked app', async () => {
